+++ 1.21.11/Fabric/src/main/java/fuzs/puzzleslib/fabric/api/event/v1/FabricLivingEvents.java
@@ -106,7 +106,8 @@ public final class FabricLivingEvents {
      * Called when a {@link MobEffectInstance} is removed from a {@link LivingEntity} in
      * <code>net.minecraft.world.entity.LivingEntity#tickEffects</code> due to the instance duration having run out.
      */
-    public static final Event<MobEffectEvents.Expire> MOB_EFFECT_EXPIRE = FabricEventFactory.create(MobEffectEvents.Expire.class);
+    public static final Event<MobEffectEvents.Expire> MOB_EFFECT_EXPIRE = FabricEventFactory.createResult(
+            MobEffectEvents.Expire.class);
     /**
      * Called when an entity is jumping, allows for modifying the jump height as well as preventing the jump.
      */