+++ 1.21.11/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add `GameRuleUpdatedCallback`
 - Add back `GatherEffectScreenTooltipCallback`
 
 ## [v21.11.3-1.21.11] - 2025-12-17