+++ 1.21.11/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.11.5-1.21.11] - 2026-01-03
+
+### Fixed
+
+- Fix `MobEffectEvents$Remove` running when the `MobEffectInstance` is `null`
+- Fix `MobEffectEvents$Expire` not allowing for returning an `EventResult`
+
 ## [v21.11.4-1.21.11] - 2025-12-20
 
 ### Added